requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
kafka-python==2.0.2
requests==2.31.0
urllib3==2.0.2
xmltodict==0.13.0
